refactor: common폴더 components폴더 아래에 위치하도록 변경
- 공통 컴포넌트가 담긴 common폴더 components폴더 아래에 위치하도록 변경
